requirements.txt
tensorflow>=2.10.0
opencv-python
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
albumentations
